+++ website/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.261.2](https://github.com/Paddione/Bachelorprojekt/compare/website-v1.261.1...website-v1.261.2) (2026-08-10)
+
+
+### Bug Fixes
+
+* **scripts:** ticket.sh --help/-h auf Kommando- und Optionsebene [T002843] ([#4133](https://github.com/Paddione/Bachelorprojekt/issues/4133)) ([a48aa63](https://github.com/Paddione/Bachelorprojekt/commit/a48aa63f5afbfcb7bdf93c4eb38e730e15a24359))
+
+## [1.261.1](https://github.com/Paddione/Bachelorprojekt/compare/website-v1.261.0...website-v1.261.1) (2026-08-10)
+
+
+### Bug Fixes
+
+* **factory:** factory-prep stdout-Leck zerstoert Launch-Plan des gesamten Ticks [T003269] ([#4131](https://github.com/Paddione/Bachelorprojekt/issues/4131)) ([9055fdb](https://github.com/Paddione/Bachelorprojekt/commit/9055fdb17d99b27889386d4c009ba9665e948114))
+
 ## [1.261.0](https://github.com/Paddione/Bachelorprojekt/compare/website-v1.260.1...website-v1.261.0) (2026-08-10)